% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizers.R
\name{featurizers}
\alias{featurizers}
\alias{make_7dav_featurizer}
\title{Functions to create "featurizers"}
\usage{
make_7dav_featurizer()
}
\value{
A function that takes in a data frame with the same columns, 
  but possibly fewer rows.
}
\description{
Essentially, these create functions that compute
weekly averages of daily data. The function needs no arguments at the 
moment.
}
