% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_cv_idx.R
\name{forward_cv_idx}
\alias{forward_cv_idx}
\title{Helper script to determine train/validation splits for forward cross validation
currently unimplemented
the first 4 inputs are internal to the production_forecaster
nfolds / ntrain are possibly additional named params in ...
Together, these determine the validation sets for forward validation
before being passed along to cv_quantile_lasso}
\usage{
forward_cv_idx(
  time_value,
  train_end_date,
  training_window_size,
  ahead,
  nfolds,
  ntrain
)
}
\description{
Helper script to determine train/validation splits for forward cross validation
currently unimplemented
the first 4 inputs are internal to the production_forecaster
nfolds / ntrain are possibly additional named params in ...
Together, these determine the validation sets for forward validation
before being passed along to cv_quantile_lasso
}
