% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_type.R
\name{get_geo_type}
\alias{get_geo_type}
\title{Helper function to get geo_type for COVIDcast signals}
\usage{
get_geo_type(df)
}
\arguments{
\item{df}{A data.frame corresponding to a single COVIDcast signal.}
}
\value{
The `geo_type` of the signal ("state" or "county").
}
\description{
Helper function to get `geo_type` ("state" or "county") for COVIDcast 
signals. A bit of a hack: it looks at the number of characters in the 
first element in the `geo_value` column and makes a determination.
}
