% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_day.R
\name{grab_start_day}
\alias{grab_start_day}
\title{Produce start_day function}
\usage{
grab_start_day(
  ahead,
  training_window_size,
  lags,
  incidence_period,
  roll_sum_len = 7,
  cv = FALSE,
  cv_type = c("forward", "random"),
  nfolds = 5
)
}
\arguments{
\item{ahead}{vector of aheads}

\item{training_window_size}{desired size of the training set}

\item{lags}{vector (or list) of lags for features}

\item{incidence_period}{String indicating the incidence period, either
"epiweek" or "day".}

\item{roll_sum_len}{if features need trailing sums, how far back do we need?
Default is 7}

\item{cv}{will we be performing cv? (currently ignored)}

\item{cv_type}{forward cv or random, forward, requires more historical data
(currently ignored)}

\item{nfolds}{how many folds to use for CV (currently ignored)}
}
\value{
returns a function with a single argument: the forecast_date as a 
  string or Date object
}
\description{
Produce start_day function
}
