% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invnorm_by_population.R
\name{invnorm_by_population}
\alias{invnorm_by_population}
\title{Helper function for undoing normalization by population}
\usage{
invnorm_by_population(df, geo_type, base = 1e+05)
}
\arguments{
\item{df}{Data frame. Must include `geo_value` and `value` columns.}

\item{geo_type}{The geography type for which to request this data, either
"county" or "state".}

\item{base}{The base for the normalization, i.e. the resulting values 
should be interpreted as `value` per `base`. Default is `1e5`.}
}
\value{
`df` but with the `value` column appropriately unnormalized.
}
\description{
Helper function to undo normalization by population. This is useful when
we have made predictions for a response on a population-normalized basis,
and we want to convert them to predictions for the unnormalized response.
The function replaces the `value` column in the predictions data frame with
the unnormalized value. Currently works with "state" and "county" 
`geo_type` only.
}
