% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_processor.R
\name{lag_processor}
\alias{lag_processor}
\title{Helper function for processing lags for production forecaster}
\usage{
lag_processor(lags, nsigs)
}
\arguments{
\item{lags}{Vector of lag values to use as features in the autoregressive
model. For example, setting `lags = c(0, 7, 14)` means we use the 
current value of each signal as well as the values 7 and 14 days ago, as the
features. Note that `lags` can also be a list of vectors of lag
values, this list having the same length as the number of signals,
in order to apply a different set of shifts to each signal.}

\item{nsigs}{Number of signals.}
}
\value{
A list of length `nsigs`, with each element being the vector of
(non-positive) lags associated with each signal.
}
\description{
Helper function to take positive input lags (in the style of Ryan's
`quantgen_forecaster`), convert to a list if necessary, and do some error
checking.
}
