% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{selectors}
\alias{selectors}
\alias{select_geo_top_n}
\title{Functions to create "selectors"}
\usage{
select_geo_top_n(
  response_data_source = "jhu-csse",
  response_signal = "confirmed_incidence_num",
  n_locations = 200
)
}
\arguments{
\item{response_data_source}{The name of the response.}

\item{response_signal}{The name of the response signal.}

\item{n_locations}{The number of locations to predict.}
}
\value{
A function that takes in a data frame in wide format and returns
  a vector of geo_values.
}
\description{
These create functions that take in a dataframe in wide format and
return a vector of geo_values. These geo_values are the ones that we
want forecasts for. (We currently only use them for county case 
predictions.)
}
