% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_population}
\alias{state_population}
\title{State population data}
\format{
Data frame with 53 rows (including one for the United States as a
  whole, plus the District of Columbia and the Puerto Rico Commonwealth).

\describe{
  \item{geo_value}{2 letter abbreviation of the state / territory}
  \item{population}{Estimate of the location's resident population in 2019.}
}
}
\source{
United States Census Bureau, at
  \url{https://www2.census.gov/programs-surveys/popest/datasets/2010-2019/state/detail/SCPRC-EST2019-18+POP-RES.csv}
}
\usage{
state_population
}
\description{
Data set on state populations, from the 2019 US Census.
}
\seealso{
[covidcast::state_census]
}
\keyword{datasets}
