% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizers.R
\name{states_featurizer}
\alias{states_featurizer}
\alias{counties_featurizer}
\title{Get weekly averages of daily data}
\usage{
states_featurizer(df)

counties_featurizer(df)
}
\arguments{
\item{df}{a "wide" data frame as created by [covidcast::aggregate_signals()].
should contain the columns `geo_value` and `time_value` as well as columns
containing leads and lags. These will all begin with the string "value"}
}
\value{
a data frame with the same columns, but possibly fewer rows
}
\description{
Get weekly averages of daily data
}
\section{Functions}{
\itemize{
\item \code{counties_featurizer}: Selects top 200 locations by total cases before averaging
}}

