% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transformer}
\alias{transformer}
\title{Helper function for univariate transformations of COVIDcast signals}
\usage{
transformer(df, transform, inv_trans)
}
\arguments{
\item{df}{List of signal values to use for forecasting, of the format
that is returned by [covidcast::covidcast_signals()].}

\item{transform, inv_trans}{Transformation and inverse transformations to use
for the response/features. The former `transform` can be a function or a
list of functions, this list having the same length as the number of elements
in the `df` list, in order to apply the same transformation or a
different transformation to each signal. The latter argument `inv_trans`
specifies the inverse transformation to use on the response variable
(inverse of `transform` if this is a function, or of `transform[[1]]` if
`transform` is a list). Default is `NULL` for both `transform` and
`inv_trans`, which means no transformations are applied.}
}
\value{
`df` but with the `value` columns appropriately transformed.
}
\description{
Helper function for univariate transformations of COVIDcast signals.
}
