% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizers.R
\name{featurizers}
\alias{featurizers}
\alias{make_7dav_featurizer}
\alias{make_county_7dav_featurizer}
\title{Functions to create "featurizers"}
\usage{
make_7dav_featurizer()

make_county_7dav_featurizer(
  response_data_source = "jhu-csse",
  response_signal = "confirmed_incidence_num",
  n_locations = 200,
  location_size_filter_lookback = NULL,
  forecast_date = NULL
)
}
\arguments{
\item{response_data_source}{the name of the response}

\item{response_signal}{the name of the response signal}

\item{n_locations}{the number of locations to predict}

\item{location_size_filter_lookback}{when summing county totals to determine
the top \code{n_locations} counties, how many days back should we sum over?
Defaults to all days available. If you specify this argument, you must also
pass the forecast date.}

\item{forecast_date}{required for non null
\code{location_size_filter_lookback} otherwise ignored}
}
\value{
A function that takes in a data frame with the same columns, 
  but possibly fewer rows.
}
\description{
Essentially, these both create functions that compute
weekly averages of daily data. The state version needs no arguments at the 
moment. For counties, we need the name of the response so that we can choose
the top n counties.
}
