% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_by_population.R
\name{normalize_by_population}
\alias{normalize_by_population}
\title{Helper function for normalizing COVIDcast signals by population}
\usage{
normalize_by_population(df_list, norm_by_popn, base = 1e+05)
}
\arguments{
\item{df_list}{List of signal values to use for forecasting, of the format
that is returned by [covidcast::covidcast_signals()].}

\item{norm_by_popn}{Should the response/features be normalized by population?
This can be a single boolean value or a list of boolean values having the
same length as the number of elements in the `df` list, which tells us
which signals in `df` should be normalized by population and which
should not. Default is `FALSE`, i.e. no normalization.}

\item{base}{The base for the normalization, i.e. the resulting values 
should be interpreted as `value` per `base`. Default is `1e5`.}
}
\value{
`df_list` but with the `value` column appropriately normalized.
}
\description{
Helper function to normalize COVIDcast signals by population. This replaces
the `value` column in each signal with the appropriately normalized
value. The user can choose which signals to normalize. Currently works with
"state" and "county" `geo_type` only.
}
