% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{selectors}
\alias{selectors}
\alias{select_geo_top_n}
\title{Function factory for creating "selector" functions based on top response 
count}
\usage{
select_geo_top_n(n_locations = 200)
}
\arguments{
\item{n_locations}{The number of locations to predict/return.}
}
\value{
A function that takes in a data frame in list format and returns
  a vector of geo_values.
}
\description{
This creates a function that takes in a dataframe in list format and
returns a vector of geo_values. These geo_values are the ones that have the
largest cumulative value of the response, and are the ones that we
want forecasts for. (We currently only use them for county case 
predictions.)
}
